# apps/web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tour Reel Studio",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}": {
      "get": {
        "summary": "Get Project",
        "operationId": "get_project_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/videos": {
      "post": {
        "summary": "Create Video",
        "operationId": "create_video_projects__project_id__videos_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopicIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{video_id}": {
      "get": {
        "summary": "Get Video",
        "operationId": "get_video_videos__video_id__get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{video_id}/plans/regenerate": {
      "post": {
        "summary": "Regenerate Plans",
        "operationId": "regenerate_plans_videos__video_id__plans_regenerate_post",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{video_id}/approve-plan": {
      "post": {
        "summary": "Approve Plan",
        "operationId": "approve_plan_videos__video_id__approve_plan_post",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovePlanIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{video_id}/shots/{shot_no}/regenerate": {
      "post": {
        "summary": "Regenerate Shot",
        "operationId": "regenerate_shot_videos__video_id__shots__shot_no__regenerate_post",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          },
          {
            "name": "shot_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shot No"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RegenerateIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{video_id}/approve": {
      "post": {
        "summary": "Approve Final",
        "operationId": "approve_final_videos__video_id__approve_post",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{video_id}/download": {
      "get": {
        "summary": "Download",
        "operationId": "download_videos__video_id__download_get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{video_id}/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_videos__video_id__events_get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/higgsfield": {
      "post": {
        "summary": "Higgsfield Webhook",
        "operationId": "higgsfield_webhook_webhooks_higgsfield_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApprovePlanIn": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "cost_cap": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Cost Cap"
          },
          "placements": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/PlacementIn"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placements"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan_id",
          "cost_cap"
        ],
        "title": "ApprovePlanIn"
      },
      "BrandOut": {
        "properties": {
          "visual": {
            "additionalProperties": true,
            "type": "object",
            "title": "Visual"
          },
          "selling_points": {
            "items": {},
            "type": "array",
            "title": "Selling Points"
          },
          "tone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tone"
          },
          "info": {
            "additionalProperties": true,
            "type": "object",
            "title": "Info"
          }
        },
        "type": "object",
        "required": [
          "visual",
          "selling_points",
          "info"
        ],
        "title": "BrandOut"
      },
      "CharacterOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "anchor_card": {
            "title": "Anchor Card"
          },
          "cutout_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cutout Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "version"
        ],
        "title": "CharacterOut",
        "description": "鎖定的角色版本；cutout_url 為去背圖的預簽名網址（spec 0015 R-007）。"
      },
      "DownloadOut": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "url",
          "expires_at"
        ],
        "title": "DownloadOut"
      },
      "EstimateOut": {
        "properties": {
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total"
          },
          "reserve": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Reserve"
          },
          "cap": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cap"
          }
        },
        "type": "object",
        "required": [
          "total",
          "reserve",
          "cap"
        ],
        "title": "EstimateOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PlacementIn": {
        "properties": {
          "x": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X"
          },
          "y": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y"
          },
          "scale": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Scale"
          },
          "flip": {
            "type": "boolean",
            "title": "Flip",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "x",
          "y",
          "scale"
        ],
        "title": "PlacementIn",
        "description": "角色擺放：以照片座標的比例值 0–1 表示（架構書 §5.3）。"
      },
      "PlanOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "estimate": {
            "$ref": "#/components/schemas/EstimateOut"
          }
        },
        "type": "object",
        "required": [
          "id",
          "payload",
          "estimate"
        ],
        "title": "PlanOut"
      },
      "ProjectOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "brand": {
            "$ref": "#/components/schemas/BrandOut"
          },
          "character": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CharacterOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "scene_photos": {
            "items": {
              "$ref": "#/components/schemas/ScenePhotoOut"
            },
            "type": "array",
            "title": "Scene Photos"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "brand",
          "character",
          "scene_photos"
        ],
        "title": "ProjectOut"
      },
      "RegenerateIn": {
        "properties": {
          "cost_cap": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Cap"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RegenerateIn",
        "description": "重生單鏡；cost_cap 為使用者同意提高後的成本上限（spec 0016 R-007）。"
      },
      "ScenePhotoOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "image_key": {
            "type": "string",
            "title": "Image Key"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "image_key",
          "width",
          "height"
        ],
        "title": "ScenePhotoOut"
      },
      "ShotOut": {
        "properties": {
          "shot_no": {
            "type": "integer",
            "title": "Shot No"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "duration_s": {
            "type": "number",
            "title": "Duration S"
          },
          "placement": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placement"
          },
          "take": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TakeOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "regen_cost": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Regen Cost",
            "default": "0"
          }
        },
        "type": "object",
        "required": [
          "shot_no",
          "role",
          "duration_s",
          "placement",
          "take"
        ],
        "title": "ShotOut"
      },
      "TakeOut": {
        "properties": {
          "attempt": {
            "type": "integer",
            "title": "Attempt"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "keyframe_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyframe Key"
          },
          "clip_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clip Key"
          }
        },
        "type": "object",
        "required": [
          "attempt",
          "status",
          "keyframe_key",
          "clip_key"
        ],
        "title": "TakeOut"
      },
      "TopicIn": {
        "properties": {
          "topic": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Topic"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "topic"
        ],
        "title": "TopicIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VideoOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "plans": {
            "items": {
              "$ref": "#/components/schemas/PlanOut"
            },
            "type": "array",
            "title": "Plans"
          },
          "selected_plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Plan Id"
          },
          "shots": {
            "items": {
              "$ref": "#/components/schemas/ShotOut"
            },
            "type": "array",
            "title": "Shots"
          },
          "cost_cap": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Cap"
          },
          "spent": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Spent"
          },
          "preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Url"
          },
          "fallback_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "status",
          "topic",
          "plans",
          "selected_plan_id",
          "shots",
          "cost_cap",
          "spent"
        ],
        "title": "VideoOut"
      }
    }
  }
}